requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.5.0
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.6.3
tzdata==2023.3
